debugged

+++ C_Conversion/kalman.h
@@ -47,13 +47,13 @@ void calculatePredictedVar(size_t state_size, float displacement[3][1], float va
 
 void setPositionOfNeverSeenLandmark(size_t state_size, float pred_state[state_size][1], struct Seen_Land_List seen_list, struct Observations observation);
 
-void obtainExpectedObservation(size_t state_size, float pred_state[state_size][1], int x_coor, int y_coor, float delta[2], float exp_dis_ang[2], float* q);
+void obtainExpectedObservation(size_t state_size, float pred_state[state_size][1], float x_coor, float y_coor, float delta[2], float exp_dis_ang[2], float* q);
 
 void computeJacobianHMatrix(size_t state_size, float jacobian[2][state_size], int seenLandmarkIndex, float q, float delta[2]);
 
 void computeKalmanGain(size_t state_size, float pred_var[state_size][state_size], float jacobian[2][state_size], float kalman_gain[state_size][2]);
 
-void getPredictedStateFromKalmanGain(size_t state_size, int land_dist, int land_ang, float kalman_gain[state_size][2], float exp_dis_ang[2], float pred_state[state_size][1]);
+void getPredictedStateFromKalmanGain(size_t state_size, float land_dist, float land_ang, float kalman_gain[state_size][2], float exp_dis_ang[2], float pred_state[state_size][1]);
 
 void getPredictedVarFromKalmanGain(size_t state_size, float kalman_gain[state_size][2], float jacobian[2][state_size], float pred_var[state_size][state_size]);
